backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
psutil==5.9.6
py-machineid
aiohttp==3.9.1